prompthack_suite/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
azure-ai-inference==1.0.0b4
azure-core==1.31.0
boto3==1.35.31
botocore==1.35.31
certifi==2024.6.2
charset-normalizer==3.3.2
cloudpickle==2.2.1
cohere==5.11.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dill==0.3.9
distro==1.9.0
docker==7.1.0
fastavro==1.9.7
filelock==3.13.1
fonttools==4.54.1
fsspec==2024.2.0
google-pasta==0.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
idna==3.7
importlib-metadata==6.11.0
install==1.3.5
intel-openmp==2021.4.0
isodate==0.6.1
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mkl==2021.4.0
mock==4.0.3
mpmath==1.3.0
multiprocess==0.70.17
networkx==3.2.1
numpy==1.26.3
openai==1.34.0
packaging==24.1
pandas==2.2.3
parameterized==0.9.0
pathos==0.3.3
pillow==10.2.0
platformdirs==4.3.6
pox==0.3.5
ppft==1.7.6.9
protobuf==4.25.5
psutil==6.0.0
pydantic==2.7.4
pydantic_core==2.18.4
pyfiglet==1.0.2
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==306
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.9.1
rpds-py==0.20.0
s3transfer==0.10.2
safetensors==0.4.3
sagemaker==2.232.1
sagemaker-core==1.0.9
schema==0.7.7
seaborn==0.13.2
six==1.16.0
smdebug-rulesconfig==1.0.1
sniffio==1.3.1
sympy==1.12
tbb==2021.11.0
tblib==3.0.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.3.1+cpu
torchvision==0.18.1+cpu
tqdm==4.66.4
transformers==4.41.2
types-requests==2.32.0.20240914
typing_extensions==4.9.0
tzdata==2024.2
urllib3==2.2.2
zipp==3.20.2